requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
boto3==1.34.14
python-multipart==0.0.6
pyjwt==2.8.0
passlib==1.7.4
openai==1.10.0
pypdf==4.0.1
ebooklib==0.18
beautifulsoup4==4.12.3
python-docx==1.1.0
weasyprint==60.2
jinja2==3.1.3
email-validator==2.1.0.post1